Feat: Add'l CSS → Marked headings
Add more colors for the DAESA-styled right-hand notched mark design to «CSS/wsuwp-wds_cust5r-addl-css-src.css» for deployment to the DAESA website's Additional CSS panel of the theme customizer. Also fix the style selector for changing styling of table headers in block tables so it is the default on DAESA websites.